SDK/requirement.txt
azure-mgmt-resource
azure-mgmt-web
azure-identity
